gandiv5: rename provider and enhance error messages
