+++ providers/dns/gandiv5/gandiv5.go
@@ -1,6 +1,6 @@
-// Package gandibeta implements a DNS provider for solving the DNS-01
-// challenge using Gandi DNS.
-package gandibeta
+// Package gandiv5 implements a DNS provider for solving the DNS-01
+// challenge using Gandi LiveDNS api.
+package gandiv5
 
 import (
 	"bytes"
@@ -20,7 +20,7 @@ import (
 var (
 	// endpoint is the Gandi API endpoint used by Present and
 	// CleanUp. It is overridden during tests.
-	endpoint = "https://dns.beta.gandi.net/api/v5"
+	endpoint = "https://dns.api.gandi.net/api/v5"
 	// findZoneByFqdn determines the DNS zone of an fqdn. It is overridden
 	// during tests.
 	findZoneByFqdn = acme.FindZoneByFqdn
@@ -42,9 +42,9 @@ type DNSProvider struct {
 }
 
 // NewDNSProvider returns a DNSProvider instance configured for Gandi.
-// Credentials must be passed in the environment variable: GANDI_API_KEY.
+// Credentials must be passed in the environment variable: GANDIV5_API_KEY.
 func NewDNSProvider() (*DNSProvider, error) {
-	apiKey := os.Getenv("GANDI_API_KEY")
+	apiKey := os.Getenv("GANDIV5_API_KEY")
 	return NewDNSProviderCredentials(apiKey)
 }
 
@@ -52,7 +52,7 @@ func NewDNSProvider() (*DNSProvider, error) {
 // DNSProvider instance configured for Gandi.
 func NewDNSProviderCredentials(apiKey string) (*DNSProvider, error) {
 	if apiKey == "" {
-		return nil, fmt.Errorf("No Gandi API Key given")
+		return nil, fmt.Errorf("Gandi DNS: No Gandi API Key given")
 	}
 	return &DNSProvider{
 		apiKey:          apiKey,
@@ -171,18 +171,13 @@ func (d *DNSProvider) sendRequest(method string, resource string, payload interf
 	defer resp.Body.Close()
 
 	if resp.StatusCode >= 400 {
-		return nil, fmt.Errorf("Gandi API request failed with HTTP status code %d", resp.StatusCode)
+		return nil, fmt.Errorf("Gandi DNS: request failed with HTTP status code %d", resp.StatusCode)
 	}
 	var response responseStruct
 	err = json.NewDecoder(resp.Body).Decode(&response)
 	if err != nil && method != "DELETE" {
 		return nil, err
 	}
-
-	// if dynRes.Status == "failure" {
-	// 	// TODO add better error handling
-	// 	return nil, fmt.Errorf("Dyn API request failed: %s", dynRes.Messages)
-	// }
 
 	return &response, nil
 }
@@ -196,7 +191,7 @@ func (d *DNSProvider) addTXTRecord(domain string, name string, value string, ttl
 		RRSetValues: []string{value},
 	})
 	if response != nil {
-		fmt.Printf("Gandi API answered: %s\n", response.Message)
+		fmt.Printf("Gandi DNS: %s\n", response.Message)
 	}
 	return err
 }
@@ -206,8 +201,8 @@ func (d *DNSProvider) deleteTXTRecord(domain string, name string) error {
 	response, err := d.sendRequest("DELETE", target, deleteFieldRequest{
 		Delete: true,
 	})
-	if response != nil {
-		fmt.Printf("Gandi API answered: %s\n", response.Message)
+	if response != nil && response.Message == "" {
+		fmt.Printf("Gandi DNS: Zone record deleted\n")
 	}
 	return err
 }
